Send comments with some context
This patch adds new property `linesWithContext` on comments available to
Soy templates.

The existing `lines` property only contains the line directly targeted
by a comment.  For a comment attached to a single line (most comments
are like that), `lines` will contain that single line.

The property `linesWithContext` contains the targeted line, but also
some lines before.  It includes at least 10 lines, including the
targeted line itself.

It also tries to make sure that this context includes at least one line
that could give the user a clue of where we're in.  For now, it assumes
that everything is C code.  It searches for a line that starts with a
non-whitespace character that isn't just "{".  If the 10 lines don't
already contain such a meaningful line, it will keep searching backwards
until it finds one, and put it at the top, with an ellipsis.  In the
end, it can look like this, where the comment was put on line 1006:

   983 | bt_self_component_port_input_message_iterator_can_seek_ns_from_origin(
       | ...
   997 |
   998 | 	if (iterator->methods.can_seek_ns_from_origin) {
   999 | 		/*
  1000 | 		 * Initialize to an invalid value, so we can post-assert that
  1001 | 		 * the method returned a valid value.
  1002 | 		 */
  1003 | 		*can_seek = -1;
  1004 |
  1005 | 		BT_LIB_LOGD("Calling user's \"can seek nanoseconds from origin\" method: %!+i",
  1006 | 			iterator);

If the comment is applied on a range, it also makes sure to quote at
least the whole range.

Change-Id: I204c05f5bf80dec7a94f4b4cde88600b7db9f4bc