Add script to generate diff interleaved with comments
The goal of this patch is to be able to generate comment notification
emails that show comments in a diff (like one would do when replying to
patches in a standard email workflow).  Today, we show a bit of context
around comments, but that context only includes lines from the version
on the file on which you are commenting.  This means that if the user
comments in the left pane, the comment will be shown with lines from the
"before" version of the file, whereas if they comment in the right pane,
the comment will be shown with context from the "after" version of the
file.  People would like instead to see comments in a diff context that
shows both removals and additions.

The goal of this patch is a bit to reproduce in an email what can be
seen if you look at a patch on the web UI in the "unified diff" mode.

I started to prototype this by writing a Python script using Gerrit's
REST API.  In the end, I decided to keep it as a Python script called
from Gerrit.  This seems a bit crazy, but I think it will be more
maintainable and more easily testable this way.  Also, anybody can
easily try and modify the Python script.

So the generate-comment-diff.py can be used to point a given group of
comments (which were all posted in one shot) on a given change of a
given server.  See the help of the script to know how to run it.

I have made it so that the script is automatically installed in bin/
when initializing a new site path, so it should just work out of the box
(tm) for a new installation.

Change-Id: Iad3d31e4a51bd1aeb4c8dcd2f306bc2816a14b79