Change leading character for diff lines
The diff lines are currently prefixed by the `>` character, to make it
look like it is part of an email reply.  This doesn't work / is
confusing for replies to comments, because it looks like the comment
being replied to is the who containing the diff.

Change the leading character to `|` so it doesn't look like a quote.

Change-Id: I5e0d72ed8365ebe515dbda5b0a7f58926b47e71f